OV-31: * move markers to separate file
